api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Ilnar Gomelyanov <hibushland@gmail.com>", "Ramil Suleimanov <lllumineux@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
fastapi = "^0.75.0"
aiohttp = "^3.8.1"
python-dotenv = "^0.19.2"
fastapi-utils = "^0.2.1"
SQLAlchemy = "^1.4.32"
databases = "^0.5.5"
asyncpg = "^0.25.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
